Added Model to Force Negative Values
Custom model for fields that the Balance Sheet model requires to be reflected as negative values. This approach will keep the conversion logic within the model itself. Those field within the BalanceSheet model will then reference the NegativeDecimalField model.

Please note that the conversion to a negative value will not automatically be reflected in the form until saved, unless client-side JavaScript code is included to handle this negative conversion in real-time.